google_function/gemini.py
import google.generativeai as genai
import os
import json

def googleGemini(prompt,
            model="gemini-ultra",  # Hypothetical model name
            max_tokens=4096,
            temperature=0.7,
            system_prompt="",
            conversation_history=[] # ... add other hypothetical Gemini-specific parameters
            ):
    
    # Load your Google Gemini credentials (mechanism will depend on your setup)
    #gemini_api_key = os.environ.get('GEMINI_API_KEY')  # Adjust if needed 
    #generative_model = genai.GenerativeModel(model=model)
    
    model = genai.GenerativeModel(model)
    chat = model.start_chat(history=[])
    
    message_sequence = [
       # System prompt structured with 'parts' and 'role'
       {
           "parts": [{"text": system_prompt}],
           "role": "user"
       },
       # User prompt structured with 'parts' and 'role'
       {
           "parts": [{"text": "Understood. Please provide the article."}],
           "role": "model"
       },
       
   ]
    
    chat.history = chat.history + message_sequence
    
    print(chat)
    
    generation_config=genai.types.GenerationConfig(
        max_output_tokens=max_tokens,
        temperature=temperature
    )
    # Hypothetical interaction with the Gemini model 
    try:
        response = chat.send_message(prompt)

        
        

        # Hypothetical extraction of response content.  
        # Tailor this  based on the actual response format from Gemini,
        structured_response = {
            "text": response.text, 
            "function_calls": []  # Adjust if Gemini allows function calls like GPT-4
        } 

        return structured_response

    except Exception as e:
        return {"error": str(e)} 

# ===============================
# Example Usage
# ===============================
