requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
alabaster=0.7.16=pyhd8ed1ab_0
annotated-types=0.6.0=pypi_0
anthropic=0.19.1=pypi_0
anyio=4.3.0=pypi_0
applaunchservices=0.3.0=pyhd8ed1ab_2
appnope=0.1.4=pyhd8ed1ab_0
arrow=1.3.0=pyhd8ed1ab_0
astroid=2.14.2=py312hca03da5_0
asttokens=2.4.1=pyhd8ed1ab_0
atomicwrites=1.4.1=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
autopep8=2.0.4=pyhd8ed1ab_0
babel=2.14.0=pyhd8ed1ab_0
beautifulsoup4=4.12.3=pyha770c72_0
binaryornot=0.4.4=py_1
black=24.2.0=py312h81bd7bf_0
bleach=6.1.0=pyhd8ed1ab_0
brotli-python=1.1.0=py312h9f69965_1
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.2.2=hf0a4a13_0
certifi=2024.2.2=pyhd8ed1ab_0
chardet=5.2.0=py312h81bd7bf_1
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
cloudpickle=3.0.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.1=pyhd8ed1ab_0
cookiecutter=2.6.0=pyhca7485f_0
debugpy=1.8.1=py312h20a0b95_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
diff-match-patch=20230430=pyhd8ed1ab_0
dill=0.3.8=pyhd8ed1ab_0
distro=1.9.0=pypi_0
docstring-to-markdown=0.15=pyhd8ed1ab_0
docutils=0.20.1=py312h81bd7bf_3
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=pypi_0
flake8=7.0.0=pyhd8ed1ab_0
fsspec=2024.2.0=pypi_0
gettext=0.21.1=h0186832_0
glib=2.78.4=h1059232_4
glib-tools=2.78.4=h1059232_4
gst-plugins-base=1.14.1=h313beb8_1
gstreamer=1.14.1=h80987f9_1
h11=0.14.0=pypi_0
httpcore=1.0.4=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.21.4=pypi_0
icu=68.2=hbdafb3b_0
idna=3.6=pyhd8ed1ab_0
imagesize=1.4.1=pyhd8ed1ab_0
importlib-metadata=7.0.2=pyha770c72_0
importlib_metadata=7.0.2=hd8ed1ab_0
importlib_resources=6.1.3=pyhd8ed1ab_0
inflection=0.5.1=pyh9f0ad1d_0
intervaltree=3.1.0=pyhd8ed1ab_1
ipykernel=6.29.3=pyh3cd1d5f_0
ipython=8.22.2=pyh707e725_0
isort=5.13.2=pyhd8ed1ab_0
jaraco.classes=3.3.1=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jellyfish=1.0.3=py312h5280bc4_0
jinja2=3.1.3=pyhd8ed1ab_0
jpeg=9e=h1a8c8d9_3
jsonschema=4.21.1=pyhd8ed1ab_0
jsonschema-specifications=2023.12.1=pyhd8ed1ab_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.1=py312h81bd7bf_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_1
keyring=24.3.1=py312h81bd7bf_0
krb5=1.20.1=h69eda48_0
lazy-object-proxy=1.10.0=py312he37b823_0
libclang=12.0.0=default_hc321e17_4
libcxx=16.0.6=h4653b0c_0
libedit=3.1.20191231=hc8eb9b7_2
libexpat=2.6.1=hebf3989_0
libffi=3.4.2=h3422bc3_5
libglib=2.78.4=hfc324ee_4
libiconv=1.17=h0d3ecfb_2
libllvm12=12.0.1=h93073aa_2
libpng=1.6.43=h091b4b1_0
libpq=12.17=h02f6b3c_0
libsodium=1.0.18=h27ca646_1
libspatialindex=1.9.3=hbdafb3b_4
libsqlite=3.45.1=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.5=py312he37b823_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mccabe=0.7.0=pyhd8ed1ab_0
mdurl=0.1.2=pyhd8ed1ab_0
mistune=3.0.2=pyhd8ed1ab_0
more-itertools=10.2.0=pyhd8ed1ab_0
mypy_extensions=1.0.0=pyha770c72_0
nbclient=0.8.0=pyhd8ed1ab_0
nbconvert=7.16.2=pyhd8ed1ab_0
nbconvert-core=7.16.2=pyhd8ed1ab_0
nbconvert-pandoc=7.16.2=pyhd8ed1ab_0
nbformat=5.9.2=pyhd8ed1ab_0
ncurses=6.4=h463b476_2
nest-asyncio=1.6.0=pyhd8ed1ab_0
nspr=4.35=hb7217d7_0
nss=3.98=h5ce2875_0
numpy=1.26.4=pypi_0
numpydoc=1.6.0=pyhd8ed1ab_0
openai=1.13.3=pypi_0
openssl=3.2.1=h0d3ecfb_0
packaging=23.2=pyhd8ed1ab_0
pandoc=3.1.12.2=hce30654_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.12.1=pyhd8ed1ab_0
pcre2=10.43=h26f9a81_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pip=24.0=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.2.0=pyhd8ed1ab_0
pluggy=1.4.0=pyhd8ed1ab_0
ply=3.11=py_1
prompt-toolkit=3.0.42=pyha770c72_0
psutil=5.9.8=py312he37b823_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycodestyle=2.11.1=pyhd8ed1ab_0
pydantic=2.6.3=pypi_0
pydantic-core=2.16.3=pypi_0
pydocstyle=6.3.0=pyhd8ed1ab_0
pyflakes=3.2.0=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pylint=2.16.2=py312hca03da5_0
pylint-venv=3.0.3=pyhd8ed1ab_0
pyls-spyder=0.4.0=pyhd8ed1ab_0
pyobjc-core=10.1=py312h9d22092_0
pyobjc-framework-cocoa=10.1=py312h9d22092_0
pyobjc-framework-coreservices=10.1=py312he37b823_0
pyobjc-framework-fsevents=10.1=py312h81bd7bf_0
pyqt=5.15.10=py312h313beb8_0
pyqt5-sip=12.13.0=py312h80987f9_0
pyqtwebengine=5.15.10=py312h313beb8_0
pysocks=1.7.1=pyha2e5f31_6
python=3.12.2=hdf0ec26_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
python-fastjsonschema=2.19.1=pyhd8ed1ab_0
python-lsp-black=2.0.0=pyhd8ed1ab_0
python-lsp-jsonrpc=1.1.2=pyhd8ed1ab_0
python-lsp-server=1.10.0=pyhd8ed1ab_0
python-lsp-server-base=1.10.0=pyhd8ed1ab_0
python-slugify=8.0.4=pyhd8ed1ab_0
python.app=1.4=py312h7fd2b6f_3
python_abi=3.12=4_cp312
pytoolconfig=1.2.5=pyhd8ed1ab_0
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py312h02f2b3b_1
pyzmq=25.1.2=py312h1edf716_0
qdarkstyle=3.2.3=pyhd8ed1ab_0
qstylizer=0.2.2=pyhd8ed1ab_0
qt-main=5.15.2=ha2d02b5_7
qt-webengine=5.15.9=h2903aaf_7
qtawesome=1.3.0=pyh9208f05_0
qtconsole=5.5.1=pyhd8ed1ab_0
qtconsole-base=5.5.1=pyha770c72_0
qtpy=2.4.1=pyhd8ed1ab_0
readline=8.2=h92ec313_1
referencing=0.33.0=pyhd8ed1ab_0
requests=2.31.0=pyhd8ed1ab_0
rich=13.7.1=pyhd8ed1ab_0
rope=1.12.0=pyhd8ed1ab_0
rpds-py=0.18.0=py312h77200ec_0
rtree=1.2.0=py312h22f7183_0
setuptools=69.1.1=pyhd8ed1ab_0
sip=6.7.12=py312h650e478_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pypi_0
snowballstemmer=2.2.0=pyhd8ed1ab_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
sphinx=7.2.6=pyhd8ed1ab_0
sphinxcontrib-applehelp=1.0.8=pyhd8ed1ab_0
sphinxcontrib-devhelp=1.0.6=pyhd8ed1ab_0
sphinxcontrib-htmlhelp=2.0.5=pyhd8ed1ab_0
sphinxcontrib-jsmath=1.0.1=pyhd8ed1ab_0
sphinxcontrib-qthelp=1.0.7=pyhd8ed1ab_0
sphinxcontrib-serializinghtml=1.1.10=pyhd8ed1ab_0
spyder=5.5.1=py312h81bd7bf_0
spyder-kernels=2.5.1=unix_pyh707e725_0
sqlite=3.45.1=hf2abe2d_0
stack_data=0.6.2=pyhd8ed1ab_0
tabulate=0.9.0=pyhd8ed1ab_1
text-unidecode=1.3=pyhd8ed1ab_1
textdistance=4.5.0=pyhd8ed1ab_0
three-merge=0.1.1=pyh9f0ad1d_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tokenizers=0.15.2=pypi_0
tomli=2.0.1=pyhd8ed1ab_0
tomlkit=0.12.4=pyha770c72_0
tornado=6.4=py312he37b823_0
tqdm=4.66.2=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
types-python-dateutil=2.8.19.20240106=pyhd8ed1ab_0
typing-extensions=4.10.0=hd8ed1ab_0
typing_extensions=4.10.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
ujson=5.9.0=py312h650e478_0
urllib3=2.2.1=pyhd8ed1ab_0
watchdog=4.0.0=py312he37b823_0
wcwidth=0.2.13=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
whatthepatch=1.0.5=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
wrapt=1.16.0=py312he37b823_0
wurlitzer=3.0.3=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
yapf=0.40.1=pyhd8ed1ab_0
zeromq=4.3.5=hebf3989_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h53f4e23_5
zstd=1.5.5=h4f39d0f_0
